+++ ui/menus/characters_menu.py
@@ -2,9 +2,6 @@
 
 from colorama import Fore, Style
 
-from core.localization import get_string
-from core.models import Character
-from core.types import LanguageCode, StringsDict
 from ui.menus import _creation_steps, _deps
 from ui.menus._common import (
     _confirm_yes_no,
@@ -15,6 +12,11 @@ from ui.menus._common import (
 )
 from ui.menus._corrupt_saves import show_corrupt_save_warnings_if_any
 from ui.menus._display import _print_characters_list
+
+Character = _deps.Character
+LanguageCode = _deps.LanguageCode
+StringsDict = _deps.StringsDict
+get_string = _deps.get_string
 
 
 def _select_character_to_delete(